lab9/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
joblib==1.3.2
python-dotenv==1.0.1
sqlalchemy==2.0.27
pytest==8.0.2
httpx==0.27.0
